+++ src/view/__about__.py
@@ -1,2 +1,2 @@
-__version__ = "1.0.0-alpha3"
+__version__ = "1.0.0-alpha4"
 __license__ = "MIT"