+++ tests/test_app.py
@@ -1,7 +1,7 @@
 from ward import test
 
-from view import new_app
-
+from view import body, new_app, query
+from typing import Union
 
 @test("responses")
 async def _():
@@ -27,6 +27,7 @@ async def _():
         res = await test.get("/")
         assert res.status == 400
         assert res.message == "error"
+
 
 @test("headers")
 async def _():
@@ -78,3 +79,95 @@ async def _():
         res = await test.get("/multi")
         assert res.message == "hello"
         assert res.status == 201
+
+
+@test("body type validation")
+async def _():
+    app = new_app()
+
+    @app.get("/")
+    @body("name", str)
+    async def index(name: str):
+        return name
+
+    @app.get("/status")
+    @body("status", int)
+    async def stat(status: int):
+        return "hello", status
+
+    @app.get("/union")
+    @body("test", bool, int)
+    async def union(test: Union[bool, int]):
+        if type(test) is bool:
+            return "1"
+        elif type(test) is int:
+            return "2"
+        else:
+            raise Exception
+
+    @app.get("/multi")
+    @body("status", int)
+    @body("name", str)
+    async def multi(status: int, name: str):
+        return name, status
+
+    async with app.test() as test:
+        assert (await test.get("/", body={"name": "hi"})).message == "hi"
+        assert (await test.get("/status", body={"status": 404})).status == 404
+        assert (
+            await test.get("/status", body={"status": "hi"})
+        ).status == 400  # noqa
+        assert (await test.get("/union", body={"test": "a"})).status == 400
+        assert (
+            await test.get("/union", body={"test": "true"})
+        ).message == "1"  # noqa
+        assert (await test.get("/union", body={"test": "2"})).message == "2"
+        res = await test.get("/multi", body={"status": 404, "name": "test"})
+        assert res.status == 404
+        assert res.message == "test"
+
+
+@test("query type validation")
+async def _():
+    app = new_app()
+
+    @app.get("/")
+    @query("name", str)
+    async def index(name: str):
+        return name
+
+    @app.get("/status")
+    @query("status", int)
+    async def stat(status: int):
+        return "hello", status
+
+    @app.get("/union")
+    @query("test", bool, int)
+    async def union(test: Union[bool, int]):
+        if type(test) is bool:
+            return "1"
+        elif type(test) is int:
+            return "2"
+        else:
+            raise Exception
+
+    @app.get("/multi")
+    @query("status", int)
+    @query("name", str)
+    async def multi(status: int, name: str):
+        return name, status
+
+    async with app.test() as test:
+        assert (await test.get("/", query={"name": "hi"})).message == "hi"
+        assert (await test.get("/status", query={"status": 404})).status == 404
+        assert (
+            await test.get("/status", query={"status": "hi"})
+        ).status == 400  # noqa
+        assert (await test.get("/union", query={"test": "a"})).status == 400
+        assert (
+            await test.get("/union", query={"test": "true"})
+        ).message == "1"  # noqa
+        assert (await test.get("/union", query={"test": "2"})).message == "2"
+        res = await test.get("/multi", query={"status": 404, "name": "test"})
+        assert res.status == 404
+        assert res.message == "test"